Refactor project selector loading
